move all test code to gtest
